New slime models and anims.  Fix for new basemod/MTS (not working, pushing for review)
All 8 slime minions now use the normal in-game skeletons rather than cut sprites, and are fully animated, complete with squashing, stretching, blinking, weird eye movements...

All slime minions animate much smoother when attacking and the overhead 'Intent' image doesn't travel with them.

The Hat attachment is now attached to a bone on the character, so it will move around with the idle animation instead of being in a static position.

Fix required to work with latest MTS/Basemod.