Thematic slime positioning! (potential breakage)
Slime slots now appear on the floor in front of the Slimebound character in static positions rather than floating overhead like Defect orbs.

To make room, the Slimebound character is now pushed a little to the left.

The Slimebound's buff / debuff bar now appears above the character instead of below it, so it doesn't overlap the slimes in the case of many buffs being active.

Slime slots when empty now appear green.

Hack - The City scene's foreground gets pushed to the left to not obscure the Slimebound character.

Hack - The Awakened one fight, the Cultists get pushed closer together to make room for the Slime slots.

Hack - The Act 4 elite duo are pushed further to the sides of the screen and the Slimebound is repositioned.

Fixed Tiny Bowler Hat silhouette not being flipped.