Bug fixes and Growth Punch nerf
Growth Punch now starts at 6 damage but does not provide Block (which still starts at 1) until upgraded.  It only gains +1 to each when used, even when upgraded.  This card was just too 'swing' in terms of value, making it an instapick in Act 1 and an instant target for first upgrade, and could get into bonkers territory with the +2 upgrade (50+ damage/block for 1, seems fine...) .  It probably should still be picked as early as possible, but buffing up the base damage to at least equal a Strike may make it more pickable as the floors progress.  It should still be a high upgrade target, but isn't as immediately necessary, not until you really start needing it to provide Block.

Fixed orbs moving when the player avatar moved.

Fixed Duplicated Form causing orbs to start spawning in staggered, offset locations

Fixed (???) Licking Slimes losing their purple text... really tired of this bug.