Nerfs, purple border on goop exploits
Aesthetics

Cards that trigger bonus effects when targeting Gooped enemies (such as Leeching Touch, It Looks Tasty, etc) will now glow purple in your hand whenever an enemy is currently Gooped, helping to quickly spot a Goop exploit synergy if one is available.

Balance

Slimebound's Max HP reduced to 70. I overestimated how much max HP adjustment was needed with the previous balance update.

Leech Energy now costs 1 energy, and when hitting a Gooped enemy, restores 1 energy and draws 1 card. This puts it into a more comparable and recognizable position with Heel Hook or Dropkick, and adds a bit more Goop-exploiting focus into the kit rather than it being somewhat split between Goop exploits and Poison exploits. Only Toxin Strike exists as a Poison exploit and is solely focused on it.

Level Up now grants 1 Potency. This is a pretty big hit to the power, but as it stood in most scenarios, this was an extra 8 damage per turn guaranteed, on average doubling the base per-turn damage output of a set of 4 basic slimes - without counting bonus Slime attacks from Lead by Example, Pile On, Teamwork, or Press the Attack, or any additional slimes beyond the starting 4. It was just too much output for a single card on its own and swung deck power too heavily if one was acquired early.

Quick Study now costs 5 HP, reduced to 2 HP when upgraded. It no longer provides 2 Spire Boss cards when upgraded. This was proving to be a pretty big factor in the overall deck quality of a draft due to its upgrade essentially being two copies of the card drawn and played simultaneously - a huge boost in value for an already high-value card.

Toxin Strike now deals 5 (7*) damage, but Exhausts. The wording has been updated to not refer to 'end of turn effect', instead explaining exactly the result (lose HP equal to poison, then reduce poison by 1). The changes to this and Leech Energy now make all of the Slimebound's 0-cost cards Exhaust, which I felt was best for mechanical consistency across the kit.

Bugfixes

Fixed an issue where the 'Icky' curse could be upgraded at a rest site, crashing the game.

Fixed an issue where the Hunted event could appear if you already had acquired the Tiny Bowler Hat. This was always intended to not appear if you'd already faced the Slime Boss or acquired the Hat.