Stray+Self Forming merged, New Tackle Power, goop application tuning
[list]
[*] Stray Goop has been merged into Self-Forming Goo as a 1-energy Power that has both effects built-in - whenever you lose HP, gain 1 Block and apply 1 Goop to a random enemy at the start of your next turn.  When upgraded, the cost is reduced to 0.
[/list]

Design Note - Stray Goop and Self-Forming Goo were pretty similar in their purposes in the current iteration, so it made sense to make a merge here.

[list]
[*] Rain of Goop no longer exhausts unupgraded.  When Upgraded, now applies 2 Goop 5 times (up from 4).

[*] Corrosive Spit now applies 6 Goop (9 upgraded).

[*] Goop Spray+ now applies 14 Goop (up from 12).

[*] Douse in Slime now applies 20 Goop (30 upgraded).
[/list]

Design Note - Buffs across the board here for the non-Lick Goop applicators.  Now that these can't scale a massive amount of healing, there was room to give them a bit of a bump in offensive use by way of increased Goop amounts.

[list]
[*] Replacing Stray Goop comes a new Power, "The Best Defense".  For 1 energy, this power increases the damage your Tackles deals to enemies by 2 (3 upgraded), and reduces the damage they deal to you by 1.
[/list]

Design Note - Merging Stray and Self-Forming opened up room for a new power, and keeping in theme with trying to provide more support to Tackles, they now have a dedicated Power - a Gluttony equivalent of sorts.  This power can stack with itself and the new Relic, eventually making Tackles unable to damage you if you draft enough.  All of these changes combined brings Tackles into the forefront as a potent build path, especially for a new player who does not have access to the first tier of Tackle-related unlocks yet.